[TM-1342] move functions to clean components
