+++ src/constants/dashbordConsts.ts
@@ -1,0 +1,25 @@
+export const CHART_TYPES = {
+  multiLineChart: "multiLineChart",
+  treesPlantedBarChart: "treesPlantedBarChart"
+};
+
+export const COLORS: Record<string, string> = {
+  Total: "#053D38",
+  Enterprise: "#27A9E0",
+  "Non Profit": "#7BBD31"
+};
+
+export const MONTHS = [
+  "January",
+  "February",
+  "March",
+  "April",
+  "May",
+  "June",
+  "July",
+  "August",
+  "September",
+  "October",
+  "November",
+  "December"
+];
